chore: restructure type definitions under @types
