+++ src/client.ts
@@ -1,4 +1,4 @@
-import type { Router, InferEndpoints, Client, HTTPMethod, ClientOptions } from "@/types.ts"
+import type { Router, InferEndpoints, Client, HTTPMethod, ClientOptions } from "@/@types/index.ts"
 
 /**
  * Creates a client API for making requests to the specified router. It provides type-safe methods
