+++ src/index.ts
@@ -4,4 +4,4 @@ export { createClient } from "./client.ts"
 export { isRouterError, isInvalidZodSchemaError } from "./assert.ts"
 export { RouterError, statusCode, statusText } from "./error.ts"
 export { HeadersBuilder } from "./headers.ts"
-export type * from "./types.ts"
+export type * from "@/@types/index.ts"
